+++ CHANGELOG.md
@@ -3,9 +3,10 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.21.11-1.1.0] - 2026-01-31
 ### Added
 - Config option to display time in a 24-hour format
+- A sun/moon emoji next to the time to indicate day/night. Can be disabled in the config.
 ### Fixed
 - Negative coordinates being off by one